对敌人和墙造成伤害

+++ Assets/CShar/Bullet/Bullet.cs
@@ -9,8 +9,8 @@ public class Bullet : MonoBehaviour
     public GameObject bulletprefab;
     public Rigidbody2D bullets;
     public float speed = 5.0f;
-    public int bulletWallDamage = 10;     //对墙的子弹伤害
-    public int bulletCreatureDamage = 20; //对生物的子弹伤害
+    public float bulletWallDamage = 10f;     //对墙的子弹伤害
+    public float bulletCreatureDamage = 20f; //对生物的子弹伤害
 
     void Start()
     {
@@ -31,19 +31,20 @@ public class Bullet : MonoBehaviour
 
     private void OnCollisionEnter2D(Collision2D collision)
     {
-       Entity entity = collision.gameObject.GetComponent<Entity>();
-       if(entity != null)
-        {
-            if(entity is Wall)
-            {
-                entity.TakeDamage(bulletWallDamage);
-            }
-            if(entity is Enemy ||  entity is Player)
-            {
-                entity.TakeDamage(bulletCreatureDamage);
-            }
-            Destroy(gameObject);
-        }
+        Entity entity = collision.gameObject.GetComponent<Entity>();
+
+        if(entity != null)
+         {
+             if(entity is Wall)
+             {
+                 entity.TakeDamage(bulletWallDamage);
+             }
+             if(entity is Enemy ||  entity is Player)
+             {
+                 entity.TakeDamage(bulletCreatureDamage);
+             }
+             Destroy(gameObject);
+         }
     }
 
 }
